app/src/main/python/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
futures==3.0.5
humanfriendly==10.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pebble==5.0.3
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.0
psutil==5.9.5
py-cpuinfo==9.0.0
PyMatting==1.1.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rembg==2.0.50
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
tifffile==2023.7.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.151
urllib3==2.0.4
